requirements.txt
pytest>=8.0.0
rich>=13.0.0
openai>=1.0.0
PyYAML>=6.0.0
